refactor: remove @app/core/functions
